backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-debug-toolbar==3.2.2
django-redis-cache==3.0.0
djangorestframework==3.12.4
Faker==9.8.0
idna==3.3
isort==5.10.1
Pillow==8.4.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==3.10.0.2
urllib3==1.26.7
